♻️ Extract url functions
